Fix bug window list button lost in mobile view #68
